renamed device-connect-view-renderer to root-container-renderer and rnamed the test
