+++ src/electron/views/device-connect-view/root-container-renderer.tsx
