+++ src/tests/unit/tests/electron/views/device-connect-view/root-container-renderer.test.tsx
@@ -1,13 +1,13 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 import { BrowserWindow } from 'electron';
-import { RootContainerRenderer } from 'electron/views/device-connect-view/device-connect-view-renderer';
+import { RootContainerRenderer } from 'electron/views/device-connect-view/root-container-renderer';
 import { RootContainer, RootContainerProps } from 'electron/views/root-container/components/root-container';
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
 import { IMock, It, Mock } from 'typemoq';
 
-describe('DeviceConnectViewRendererTest', () => {
+describe('RootContainerRendererTest', () => {
     test('render', () => {
         const dom = document.createElement('div');
         const containerDiv = document.createElement('div');
@@ -19,7 +19,7 @@ describe('DeviceConnectViewRendererTest', () => {
             },
         } as BrowserWindow;
 
-        containerDiv.setAttribute('id', 'device-connect-view-container');
+        containerDiv.setAttribute('id', 'root-container');
         dom.appendChild(containerDiv);
 
         const props = {
